chore/develop-1: 폴더 구조 세팅, husky 검사 변경
