+++ .prettierrc
@@ -1,8 +1,8 @@
 {
-  "trailingComma": "es5",
-  "tabWidth": 2,
-  "semi": true,
-  "singleQuote": true,
   "printWidth": 80,
-  "bracketSpacing": true
+  "singleQuote": true,
+  "trailingComma": "none",
+  "tabWidth": 2,
+  "useTabs": false,
+  "semi": true
 }
