--- app/index.tsx
@@ -1,7 +1,0 @@
-import { Text } from 'react-native';
-
-const index = () => {
-  return <Text>index</Text>;
-};
-
-export default index;
