+++ src/app/_layout.tsx
@@ -1,0 +1,25 @@
+import {
+  DarkTheme,
+  DefaultTheme,
+  ThemeProvider
+} from '@react-navigation/native';
+import { Stack } from 'expo-router';
+import { useColorScheme } from 'react-native';
+import 'react-native-reanimated';
+import { SafeAreaProvider, SafeAreaView } from 'react-native-safe-area-context';
+
+export default function RootLayout() {
+  const colorScheme = useColorScheme();
+
+  return (
+    <SafeAreaProvider>
+      <ThemeProvider value={colorScheme === 'dark' ? DarkTheme : DefaultTheme}>
+        <SafeAreaView style={{ flex: 1 }} edges={['top', 'left', 'right']}>
+          <Stack>
+            <Stack.Screen name="index" options={{ headerShown: false }} />
+          </Stack>
+        </SafeAreaView>
+      </ThemeProvider>
+    </SafeAreaProvider>
+  );
+}
