+++ src/app/index.tsx
@@ -1,0 +1,7 @@
+import { Text } from 'react-native';
+
+const App = () => {
+  return <Text>App</Text>;
+};
+
+export default App;
