+++ src/assets/images/android-icon-foreground.png
